Reuse starlight relighting code across versions
